+++ tests/testthat/test-mixAR_diag.R
@@ -23,7 +23,12 @@ test_that("diagnostics for mixAR work",{
     expect_output( BIC_comp(list(model, model2, fit1), y) )
     
     d <- tsdiag(model, y = y, ask = FALSE)
-    expect_equal(d, tsdiag_MixARGaussian(model, y = y, ask = FALSE) )
+
+    ## this test made sense during the initial refactoring of the mixAr tsdiag method but now
+    ## the structure of the result has also changed. Some components are still as before and
+    ## could be compared individually.
+    ##
+    ## expect_equal(unclass(d), tsdiag_MixARGaussian(model, y = y, ask = FALSE) )
 
     mixAR_diag(model, y = y, ask = FALSE)
     mixAR_diag(fit1, y = y, ask = FALSE)