+++ NEWS.md
@@ -1,4 +1,8 @@
-# mixAR 0.22.4
+# mixAR 0.22.4 (CRAN)
+
+- `tsdiag` has been extensively revamped. It now is more user friendly and
+   offers more diagnostics. The object returned by it has a class and a print
+   method.
 
 - `tsdiag` now works also for models with non-Gaussian components.